Hotfix/transparency (#202)
* hotfix(main): transparency factor shouldn't exceed 1

* fix(main): revert to 10

* chore: apidump

* fix(main): add float step to Number

+++ src/main/java/cc/polyfrost/oneconfig/gui/elements/config/ConfigNumber.java
@@ -41,7 +41,7 @@ public class ConfigNumber extends BasicOption implements IFocusable {
     private boolean isFloat = true;
     private final int size;
 
-    public ConfigNumber(Field field, Object parent, String name, String description, String category, String subcategory, float min, float max, int step, int size) {
+    public ConfigNumber(Field field, Object parent, String name, String description, String category, String subcategory, float min, float max, float step, int size) {
         super(field, parent, name, description, category, subcategory, size);
         this.inputField = new NumberInputField(84, 32, 0, min, max, step);
         this.size = size;
